refactor: move variable types to clientUtils
This will allow sharing of types between server & client.
