+++ adminSiteClient/VariableEditPage.tsx
@@ -15,7 +15,7 @@ import { BindString, BindFloat, FieldsRow, Toggle } from "./Forms"
 import {
     LegacyVariableConfig,
     LegacyVariableDisplayConfig,
-} from "../grapher/core/LegacyVariableCode"
+} from "../clientUtils/OwidVariable"
 import { Grapher } from "../grapher/core/Grapher"
 import { GrapherFigureView } from "../site/GrapherFigureView"
 import { ChartList, ChartListItem } from "./ChartList"
