+++ coreTable/CoreTableColumns.ts
@@ -28,7 +28,7 @@ import { ErrorValue, ErrorValueTypes } from "./ErrorValues"
 import { getOriginalTimeColumnSlug } from "./OwidTableUtil"
 import { imemo } from "./CoreTableUtils"
 import moment from "moment"
-import { OwidSource } from "./OwidSource"
+import { OwidSource } from "../clientUtils/OwidSource"
 import { formatValue, TickFormattingOptions } from "../clientUtils/formatValue"
 import { LegacyVariableDisplayConfigInterface } from "../clientUtils/LegacyVariableDisplayConfigInterface"
 import { ColumnSlug } from "../clientUtils/owidTypes"
